Providing an example of parsing properties of Order with primitive types.
FixStringEnumerator: Provides an enumerator over a FIX string. On each enumeration it will return the next element of fix string, separated with vertical bar.
FixStringEnumerable: Provides an enumerable over a FIX string using FixStringEnumerator.
ReadonlyMemoryExtensions: Extension methods for ReadonlyMemory<char> representing a FIX element to get FIX tag and value of the element.

Added implementation of FIX string parser in FixParser. It will get the attributes from the type of TMessage type parameter, iterates through the FIX elements with FixStringEnumerable and for each element, if the element's tag had an attribute in the type of TMessage, it will set the value of the element on the TMessage instance, then return the instance.

Added editorconfig.